Time: 10 ms (10.14%), Space: 41.8 MB (7.18%) - LeetHub

+++ 0405-convert-a-number-to-hexadecimal/0405-convert-a-number-to-hexadecimal.java
@@ -11,7 +11,7 @@ class Solution {
         }  
         while(val != 0)
         {
-            long rem = (int)(val%16);
+            long rem = (val%16);
             
             if(rem>9)
             {
